pull from upstream

+++ tensorflow/core/kernels/concat_op_test.cc
@@ -1,4 +1,4 @@
-/* Copyright 2015 Google Inc. All Rights Reserved.
+/* Copyright 2015 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -28,7 +28,7 @@ limitations under the License.
 #include "tensorflow/core/kernels/ops_testutil.h"
 #include "tensorflow/core/kernels/ops_util.h"
 #include "tensorflow/core/lib/core/status_test_util.h"
-#include "tensorflow/core/platform/mem.h"
+#include "tensorflow/core/platform/prefetch.h"
 #include "tensorflow/core/platform/test.h"
 #include "tensorflow/core/platform/test_benchmark.h"
 
@@ -137,8 +137,8 @@ static void MemcpyAlternativeHelper(int iters, int concat_dimension, int dim2) {
                           ((kDim1 * dim2) + (kDim1 * dim2)) * sizeof(float));
   testing::StartTiming();
   while (--iters > 0) {
-    const int n0 = data1.size();
-    const int n1 = data2.size();
+    const size_t n0 = data1.size();
+    const size_t n1 = data2.size();
     float* result = new float[n0 + n1];
     memcpy(&result[0], &data1[0], n0 * sizeof(float));
     memcpy(&result[n0], &data2[0], n1 * sizeof(float));
@@ -157,7 +157,8 @@ BENCHMARK(BM_MemcpyAlternativeDim0)->Arg(1000)->Arg(100000)->Arg(1000000);
 BENCHMARK(BM_MemcpyAlternativeDim1)->Arg(1000)->Arg(100000)->Arg(1000000);
 
 typedef Eigen::TensorMap<Eigen::Tensor<bfloat16, 1, Eigen::RowMajor>,
-                         Eigen::Unaligned> EigenMap;
+                         Eigen::Unaligned>
+    EigenMap;
 static void MemcpyManyAlternative1(int iters, int dim2) {
   testing::StopTiming();
